fuck fix build

+++ drivers/nfc/oplus_nfc/oplus_nfc.c
@@ -88,7 +88,7 @@ static int nfc_read_func(struct seq_file *s, void *v)
 {
 	void *p = s->private;
 
-	switch((uint32_t)(p)) {
+	switch((uint64_t)(p)) {
 	case NFC_CHIPSET_VERSION:
 		seq_printf(s, "%s", current_chipset);
 		break;
